Revert "Implement rigidification in C++"
This reverts commit 095d1ae13616b5d9f90af9fb458796cd6f6518ac.
