+++ experimental/tools/elastic-miter/elastic-miter.cpp
@@ -27,7 +27,7 @@
 #include "ElasticMiterTestbench.h"
 #include "FabricGeneration.h"
 #include "GetSequenceLength.h"
-#include "experimental/Support/SmvUtils.h"
+#include "SmvUtils.h"
 
 namespace cl = llvm::cl;
 using namespace mlir;
